+++ packages/cli/src/controller/codegen-cosmos.test.ts
@@ -4,26 +4,25 @@
 import fs from 'fs';
 import path from 'path';
 import {promisify} from 'util';
-import {generateProto, tempProtoDir} from '@subql/common-cosmos';
+import {loadFromJsonOrYaml, NETWORK_FAMILY} from '@subql/common';
+import {ProjectManifestV1_0_0} from '@subql/types-core';
+import type {CosmosDatasource, CustomDatasourceTemplate, RuntimeDatasourceTemplate} from '@subql/types-cosmos';
 import {upperFirst} from 'lodash';
 import rimraf from 'rimraf';
+import {loadDependency} from '../modulars';
 import {prepareDirPath, renderTemplate} from '../utils';
 
 const PROJECT_PATH = path.join(__dirname, '../../test/protoTest1');
-const MOCK_CHAINTYPES: any = [
-  {
-    'osmosis.gamm.v1beta1': {
-      file: './proto/osmosis/gamm/v1beta1/tx.proto',
-      messages: ['MsgSwapExactAmountIn'],
-    },
+const MOCK_CHAINTYPES: any = {
+  'osmosis.gamm.v1beta1': {
+    file: './proto/osmosis/gamm/v1beta1/tx.proto',
+    messages: ['MsgSwapExactAmountIn'],
   },
-  {
-    'osmosis.poolmanager.v1beta1': {
-      file: './proto/osmosis/poolmanager/v1beta1/swap_route.proto',
-      messages: ['SwapAmountInRoute'],
-    },
+  'osmosis.poolmanager.v1beta1': {
+    file: './proto/osmosis/poolmanager/v1beta1/swap_route.proto',
+    messages: ['SwapAmountInRoute'],
   },
-];
+};
 
 jest.setTimeout(30000);
 
@@ -31,6 +30,13 @@ describe('Able to generate cosmos types from protobuf', () => {
   afterEach(async () => {
     await promisify(rimraf)(path.join(__dirname, '../../test/protoTest1/src'));
   });
+
+  const cosmosModule = loadDependency(NETWORK_FAMILY.cosmos);
+  const manifest = loadFromJsonOrYaml(path.join(PROJECT_PATH, 'project.yaml')) as ProjectManifestV1_0_0<
+    CosmosDatasource,
+    RuntimeDatasourceTemplate | CustomDatasourceTemplate,
+    any
+  >;
 
   it('Able to generate ts types from protobufs', async () => {
     const expectedGeneratedCode =
@@ -49,34 +55,26 @@ export namespace osmosis.gamm.v1beta1.tx {
 }
 
 `;
-    await generateProto(MOCK_CHAINTYPES, PROJECT_PATH, prepareDirPath, renderTemplate, upperFirst);
+
+    manifest.network.chaintypes = MOCK_CHAINTYPES;
+    await cosmosModule.projectCodegen([manifest], PROJECT_PATH, prepareDirPath, renderTemplate, upperFirst, []);
     const codegenResult = await fs.promises.readFile(path.join(PROJECT_PATH, '/src/types/CosmosMessageTypes.ts'));
     expect(fs.existsSync(`${PROJECT_PATH}/src/types/CosmosMessageTypes.ts`)).toBeTruthy();
     expect(codegenResult.toString()).toBe(expectedGeneratedCode);
   });
 
   it('On missing protobuf dependency should throw', async () => {
-    const badChainTypes = [
-      {
-        'osmosis.gamm.v1beta1': {
-          file: './proto/cosmos/osmosis/gamm/v1beta1/tx.proto',
-          messages: ['MsgSwapExactAmountIn'],
-        },
+    const badChainType = {
+      'osmosis.gamm.v1beta1': {
+        file: './proto/cosmos/osmosis/gamm/v1beta1/tx.proto',
+        messages: ['MsgSwapExactAmountIn'],
       },
-    ];
-    await expect(
-      generateProto(badChainTypes, PROJECT_PATH, prepareDirPath, renderTemplate, upperFirst)
+    };
+    manifest.network.chaintypes = badChainType;
+    await expect(() =>
+      cosmosModule.projectCodegen([manifest], PROJECT_PATH, prepareDirPath, renderTemplate, upperFirst, [])
     ).rejects.toThrow(
       'Failed to generate from protobufs. Error: chainType osmosis.gamm.v1beta1, file ./proto/cosmos/osmosis/gamm/v1beta1/tx.proto does not exist'
     );
-  });
-
-  it('create temp dir with all protobufs', async () => {
-    // user Protobufs should not be overwritten
-    const preFile = await fs.promises.readFile(path.join(PROJECT_PATH, 'proto/osmosis/gamm/v1beta1/tx.proto'));
-    const tmpDir = await tempProtoDir(PROJECT_PATH);
-    const afterFile = await fs.promises.readFile(path.join(tmpDir, 'osmosis/gamm/v1beta1/tx.proto'));
-    expect(preFile.toString()).toBe(afterFile.toString());
-    await promisify(rimraf)(tmpDir);
   });
 });