+++ packages/cli/src/controller/init-controller.ts
@@ -6,8 +6,7 @@ import childProcess, {execSync} from 'child_process';
 import fs from 'fs';
 import * as path from 'path';
 import {promisify} from 'util';
-import {DEFAULT_MANIFEST, DEFAULT_TS_MANIFEST, loadFromJsonOrYaml, makeTempDir} from '@subql/common';
-import {parseEthereumProjectManifest} from '@subql/common-ethereum';
+import {DEFAULT_MANIFEST, DEFAULT_TS_MANIFEST, loadFromJsonOrYaml, makeTempDir, NETWORK_FAMILY} from '@subql/common';
 import {ProjectManifestV1_0_0} from '@subql/types-core';
 import axios from 'axios';
 import {copySync} from 'fs-extra';
@@ -15,6 +14,7 @@ import rimraf from 'rimraf';
 import git from 'simple-git';
 import {parseDocument, YAMLMap, YAMLSeq} from 'yaml';
 import {BASE_TEMPLATE_URl, CAPTURE_CHAIN_ID_REG, CHAIN_ID_REG, ENDPOINT_REG} from '../constants';
+import {loadDependency} from '../modulars';
 import {isProjectSpecV1_0_0, ProjectSpecBase} from '../types';
 import {
   defaultEnvDevelopLocalPath,
@@ -386,7 +386,9 @@ export async function validateEthereumProjectManifest(projectPath: string): Prom
     manifest = loadFromJsonOrYaml(path.join(projectPath, DEFAULT_MANIFEST));
   }
   try {
-    return isTs ? validateEthereumTsManifest(manifest) : !!parseEthereumProjectManifest(manifest);
+    return isTs
+      ? validateEthereumTsManifest(manifest)
+      : !!loadDependency(NETWORK_FAMILY.ethereum).parseProjectManifest(manifest);
   } catch (e) {
     return false;
   }