+++ packages/cli/src/modulars/utils.ts
@@ -1,0 +1,4 @@
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// SPDX-License-Identifier: GPL-3.0
+
+import {NETWORK_FAMILY, runnerMapping} from '@subql/common';