+++ packages/cli/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - The IPFS CID for multi-chain requires the directory CID.
 
+### Changed
+- Breaking Change: Removed common network packages from dependencies, only import required module from local/global when using. (#2456)
+
 ## [4.15.0] - 2024-06-21
 ### Changed
 - Update `@subql/utils`, `@subql/common` and other dependencies